improve(api): make api a public package
