+++ packages/api/src/apps/README.md
@@ -4,7 +4,19 @@ This package is meant to ingest and expose data about the UMA ecosystem through 
 
 ## Apps
 
-- [api](./api/README.md): entry point to run all the api logic
+- [api](./apps/api): entry point to run all the api logic
+- [lsp_api](./apps/lsp_api): entry point to run all the api logic
+
+## Libraries
+
+- [libs](./libs) - various supporting libraries
+- [services](./services) - microservice style components
+- [tables](./tables) - data tables for storing api related data
+- [examples](./examples) - various frontend tools
+
+### ENV
+
+You will need a .env for each application. See the specific application readme.
 
 ## Quickstart
 
@@ -13,6 +25,6 @@ There maybe 1 or more available apps in the `src/apps` folder. To start the `api
 Install dependencies: `yarn`
 
 To run `api`:
-`npx ts-node src/start api`
 
-See each apps readme for particular environment variables needed.
+1. `yarn build`
+2. `yarn api`
