+++ packages/api/src/examples/index.ts
@@ -1,0 +1,1 @@
+export * as actionClient from "./action-client";
