+++ packages/api/src/services/multi-lsp-creator.ts
@@ -1,6 +1,6 @@
 import { clients } from "@uma/sdk";
 import bluebird from "bluebird";
-import { AppState, BaseConfig } from "..";
+import { AppState, BaseConfig } from "../types";
 import LspCreator, { EmitData, Events } from "./lsp-creator";
 
 const { lspCreator } = clients;
